python/monitor/health_check/requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
html5lib==1.0.1
PyYAML==3.13
requests==2.19.1
urllib3==1.23